use 64bit on largebytearray class and initialize binary_large_variant bool to false

+++ cpp/src/parquet/types.h
@@ -594,12 +594,12 @@ inline bool operator!=(const ByteArray& left, const ByteArray& right) {
 
 struct LargeByteArray {
   LargeByteArray() : len(0), ptr(NULLPTR) {}
-  LargeByteArray(uint32_t len, const uint8_t* ptr) : len(len), ptr(ptr) {}
+  LargeByteArray(uint64_t len, const uint8_t* ptr) : len(len), ptr(ptr) {}
 
   LargeByteArray(::std::string_view view)  // NOLINT implicit conversion
       : LargeByteArray(view.size(),
                   reinterpret_cast<const uint8_t*>(view.data())) {}
-  uint32_t len;
+  uint64_t len;
   const uint8_t* ptr;
 };
